requirements.txt

psycopg2-binary==2.8.5
pandas==1.0.5
numpy==1.19.1
requests==2.24.0
